split into tame-containers crate
