+++ crates/tame-containers/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "tame-containers"
+version = "0.1.0"
+edition = "2021"
+authors = ["Embark <opensource@embark-studios.com>", "Gray Olson <gray@grayolson.com>"]
+repository = "https://github.com/EmbarkStudios/mirror-mirror"
+homepage = "https://github.com/EmbarkStudios/mirror-mirror"
+license = "MIT OR Apache-2.0"
+readme = "../../README.md"
+rust-version = "1.65"
+description = "Containers specialized for ergonomic gamedev usecases"
+keywords = ["containers"]
+
+[features]
+default = ["speedy", "serde"]
+speedy = ["dep:speedy"]
+serde = ["dep:serde"]
+
+[dependencies]
+ahash = { version = "0.8.2", default-features = false, features = ["runtime-rng"] }
+indexmap = "1.9.3"
+once_cell = { version = "1.16", features = ["alloc", "race", "critical-section"], default-features = false }
+serde = { version = "1.0.158", default-features = false, features = ["derive", "std"], optional = true }
+speedy = { version = "0.8.5", optional = true }
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
