+++ crates/tame-containers/src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod ordered_map;
+mod unordered_map;
